Test scenarios created. Results pending. Client and server changed to use double for money.

+++ src/record.cpp
@@ -20,7 +20,7 @@ Record::Record()
 
 };
 // Record constructor
-Record::Record(int acc, std::string nm, int blnc)
+Record::Record(int acc, std::string nm, double blnc)
 {
 	account = acc;
 	name = nm;
